Engine fuly working on linux

+++ Core/Src/Render/MeshSerializer.cpp
@@ -12,24 +12,24 @@ IResource*	CMeshSerializer::Load( CDataStream& dataStream, const CResourceDescri
 
 
 	// Vertex count
-	size_t vertexCount = 0;
+	int vertexCount = 0;
 	if ( !dataStream.Read( vertexCount ) )	return NULL;
 	printf( "Vertex count %zd\n", vertexCount );
 
 	// Vertex format
 	CVertexFormat vertexFormat;
 	
-	size_t singleElementFlags = 0;
+	int singleElementFlags = 0;
 	if ( !dataStream.Read( singleElementFlags ) ) return NULL;
 	vertexFormat.SetSingleElementsFlags( (int)singleElementFlags );
 	printf( "Single elem flags %zd\n", singleElementFlags );
 	
 	for( size_t iElem = 0 ; iElem < CVertexFormat::eME_Count ; ++iElem )
 	{
-		size_t elemCount = 0;
+		int elemCount = 0;
 		if ( !dataStream.Read( elemCount ) ) return NULL;
-		vertexFormat.SetMultipleElementCount( (CVertexFormat::EMultipleElement)iElem, elemCount );
-		printf( "Multiple elem %zd count %zd\n", iElem, elemCount );
+		vertexFormat.SetMultipleElementCount( (CVertexFormat::EMultipleElement)iElem, (int)elemCount );
+		printf( "Multiple elem %zd count %d\n", iElem, elemCount );
 	}
 	
 	vertexFormat.Build();
@@ -41,18 +41,18 @@ IResource*	CMeshSerializer::Load( CDataStream& dataStream, const CResourceDescri
 	dataStream.SetCursor( dataStream.GetCursor() + vertexCount * vertexFormat.GetSize() );
 	
 	// Parts
-	size_t partCount = 0;
+	int partCount = 0;
 	if ( !dataStream.Read( partCount ) ) return NULL;
 	printf( "Part count %zd\n", partCount );
 	for( size_t iPart = 0 ; iPart < partCount ; ++iPart )
 	{
-		size_t indexCount = 0;	
+		int indexCount = 0;	
 		if ( !dataStream.Read( indexCount ) )
 		{
 			delete pMesh;
 			return NULL;
 		}
-		printf( "Index count %zd\n", indexCount );
+		printf( "Index count %d\n", indexCount );
 		
 		CMesh::CPart* pMeshPart = new CMesh::CPart( *pMesh );
 		pMeshPart->SetIndices( indexCount, (uint*)dataStream.GetCursorData() );
